+++ CHANGELOG.md
@@ -12,3 +12,7 @@
 - Text field, Text area: Added aligner utility
 - MenuItem: Added side support for the aligner
 - Button: Renamed aligned position to side, deprecated position
+- View: Static position support
+- Popover, Tooltip, DropdownMenu: Updated transition effect
+
+- Theming: Fixed setting color mode to an already applied color mode (transition)